% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dat_MstepB.R
\name{fit_dat_B}
\alias{fit_dat_B}
\title{Fit a Penalized Generalized Mixed Model via Monte Carlo Expectation Conditional 
Minimization (MCECM)}
\usage{
fit_dat_B(dat, lambda0 = 0, lambda1 = 0, conv_EM = 0.001,
  conv_CD = 1e-04, family = "binomial", offset_fit = NULL,
  trace = 0, penalty = c("MCP", "SCAD", "lasso"), alpha = 1,
  gamma_penalty = switch(penalty[1], SCAD = 4, 3),
  group_X = 0:(ncol(dat$X) - 1), nMC_burnin = 500, nMC = 5000,
  nMC_max = 20000, t = 2, nMC_report = 5000, u_init = NULL,
  coef_old = NULL, ufull_describe = NULL, maxitEM = 100,
  maxit_CD = 250, M = 10^4, sampler = c("stan", "random_walk",
  "independence"), adapt_RW_options = adaptControl(),
  covar = c("unstructured", "independent"), var_start = 0.5,
  max_cores = 1)
}
\arguments{
\item{dat}{a list object specifying y (response vector), X (model matrix of all covariates), 
Z (model matrix for the random effects), and group (vector whose value indicates 
the study, batch, or other group identity to which on observation belongs), with each row an 
observation}

\item{lambda0}{a non-negative numeric penalty parameter for the fixed effects parameters}

\item{lambda1}{a non-negative numeric penalty parameter for the grouped random effects
covariance parameters}

\item{conv_EM}{a non-negative numeric convergence criteria for the convergence of the 
EM algorithm. Default is 0.001.}

\item{conv_CD}{a non-negative numeric convergence criteria for the convergence of the 
grouped coordinate descent loop within the M step of the EM algorithm. Default 0.0001.}

\item{family}{a description of the error distribution and link function to be used in the model. 
For \code{fit_dat_B} this can be a character string naming a family function or a family function. 
See \code{family} options in the Details section.}

\item{offset_fit}{This can be used to specify an a priori known component to be included in the 
linear predictor during fitting. This should be NULL or a numeric vector of length equal to the 
number of cases.}

\item{trace}{an integer specifying print output to include as function runs. Default value is 0 
(details ...)}

\item{penalty}{character descripting the type of penalty to use in the variable selection procedure.
Options include 'MCP', 'SCAD', and 'lasso'. Default is MCP penalty.}

\item{alpha}{Tuning parameter for the Mnet estimator which controls the relative contributions 
from the MCP/SCAD/lasso penalty and the ridge, or L2, penalty. \code{alpha=1} is equivalent to 
the MCP/SCAD/lasso penalty, while \code{alpha=0} is equivalent to ridge regression. However,
\code{alpha=0} is not supported; \code{alpha} may be arbibrarily small, but not exactly zero}

\item{gamma_penalty}{The tuning parameter of the MCP and SCAD penalties. Not used by Lasso penalty.
Default is 4.0 for SCAD and 3.0 for MCP.}

\item{group_X}{vector ...}

\item{nMC}{a positive integer for the initial number of Monte Carlo draws}

\item{nMC_max}{a positive integer for the maximum number of allowed Monte Carlo draws}

\item{t}{the convergence criteria is based on the average Euclidean distance between 
the most recent coefficient estimate and the coefficient estimate from t EM iterations back.
Positive integer, default equals 2.}

\item{nMC_report}{positive integer specifying number of Monte Carlo posterior draws to return
when the function ends. Default 5,000. Warning: the returned draws are formatted in a regular
matrix (not a big.matrix). Therefore, depending on the number of random effect covariates (q)
and the number of groups (d), choose \code{nMC_report} such that a matrix of size 
\code{nMC_report} by (q*d) does not cause memory issues on your operating system.}

\item{maxitEM}{a positive integer for the maximum number of allowed EM iterations. Default equals 100.}

\item{maxit_CD}{a positive integer for the maximum number of allowed interations for the
coordinate descent algorithms used within each EM iteration. Default equals 250.}

\item{M}{positive integer specifying the number of posterior draws to use within the 
Pajor log-likelihood calculation}

\item{sampler}{character value specifying whether the Metropolis-within-Gibbs procedure 
should incorporate an adaptive random walk sampler (default, "random_walk") or an
independence sampler ("independence").}

\item{adapt_RW_options}{a list of class "adaptControl" from function \code{\link{adaptControl}} 
containing the control parameters for the adaptive random walk Metropolis-within-Gibbs procedure. 
Ignored if \code{\link{optimControl}} parameter \code{sampler} is set to "independence"}

\item{covar}{character value specifying whether the covariance matrix should be unstructured
("unstructured") or diagonal with no covariances between variables ("independent").
Default is "unstructured", but if the number of random effects (including the intercept) is 
greater than 7 (i.e. high dimensional), then the algorithm automatically assumes an 
independent covariance structure (covar switched to "independent").}

\item{fit_type}{integer specifying the verion of the grouped coordinate descent algorithm to use.
Default equals 1, the 'naive fit', assuming that the random effects covariates do not need any
additional standardization or adjustments to account for absence of orthogonalization.
For the time being, this version is recommended. Version 2 adds standardization of the random
effects covariates compared to version 1.}
}
\description{
Description
}
\section{Details}{
 
Accepted families: binomial, poisson, gaussian. Currently, the algorithm is only equipted to handle
the canonical links associated with these families (logit link for binomial, log link for poisson,
identity link for gaussian).
}

