% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmPen.R
\name{lambdaControl}
\alias{lambdaControl}
\alias{selectControl}
\title{Control of Penalized Generalized Linear Mixed Model Fitting

Constructs control structures for penalized mixed model fitting.}
\usage{
lambdaControl(lambda0 = 0, lambda1 = 0)

selectControl(lambda0_seq = NULL, lambda1_seq = NULL, nlambda = 40)
}
\arguments{
\item{lambda0}{a non-negative numeric penalty parameter for the fixed effects parameters}

\item{lambda1}{a non-negative numeric penalty parameter for the grouped random effects
covariance parameters}

\item{lambda0_seq}{a range of non-negative numeric penalty parameter for the fixed 
effects parameters. If \code{NULL}, then a range will be calculated as described by (...).}

\item{lambda1_seq}{a range of non-negative numeric penalty parameter for the grouped random 
effects covariance parameters. If \code{NULL}, then a range will be calculated as described 
by (...).}
}
\value{
The *Control functions return a list (inheriting from class "\code{pglmmControl}") 
containing penalization parameter values, presented either as an individual set or as a range of
possible values.
}
\description{
Control of Penalized Generalized Linear Mixed Model Fitting

Constructs control structures for penalized mixed model fitting.
}
