% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetClassDefn.R
\docType{class}
\name{pglmmObj-class}
\alias{pglmmObj-class}
\alias{pglmmObj}
\alias{show,}
\alias{pglmmObj-method,}
\alias{coef.pglmmObj,}
\alias{fitted.pglmmObj,}
\alias{fixef.pglmmObj,}
\alias{formula.pglmmObj,}
\alias{logLik.pglmmObj,}
\alias{model.frame.pglmmObj,}
\alias{model.matrix.pglmmObj,}
\alias{print.pglmmObj,}
\alias{ranef.pglmmObj,}
\alias{residuals.pglmmObj,}
\alias{sigma.pglmmObj,}
\alias{summary.pglmmObj}
\title{Class "pglmmObj" of Fitted Penalized Generalized Mixed-Effects Models}
\description{
The functions \code{\link{glmm}}, \code{\link{glmmPen}}, and 
\code{\link{glmmPen_FineSearch}} output the reference class object of type \code{pglmmObj}.
}

\examples{

showClass("pglmmObj")
methods(class = "pglmmObj")

}
\keyword{classes}
