% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mc_functions.R
\name{sample.mc2}
\alias{sample.mc2}
\title{Calculate Monte Carlo draws}
\usage{
sample.mc2(fit, cov, y, X, Z, nMC, trace = 0, family = family, group,
  d, nZ, okindex, gibbs = F, uold)
}
\arguments{
\item{fit}{a grpreg fit object (set \code{\link[grpreg]{grpreg}})}

\item{cov}{the random effects covariance matrix estimated from the last M step of the EM algorithm}

\item{y}{a numeric vector of the response variable}

\item{X}{a model matrix of the fixed effects covariates}

\item{Z}{a sparse model matrix of the random effects}

\item{nMC}{a positive integer for the initial number of Monte Carlo draws}

\item{trace}{an integer specifying print output to include as function runs. Default value is 0 
(details ...), with alternative options of 1 (details ...) or 2 (output acceptance rate information
of the Monte Carlo draws computed in \code{\link{sample.mc2}})}

\item{family}{a description of the error distribution and link function to be used in the model. 
For \code{fit_dat} this can be a character string naming a family function or a fmaily function. 
See \code{family} options in the Details section.}

\item{group}{a factor vector of the grouping variable, converted to a factor of 
consecutive numeric integers}

\item{d}{integer, the number of groups present (number factors of the group vector)}

\item{okindex}{?}

\item{gibbs}{logical, should Metropolis-within-Gibbs sampling be used for the Monte Carlo draws 
(if \code{TRUE}) or should Rejection sampling be used for the draws (if \code{FALSE})?}

\item{uold}{a matrix of the Monte Carlo draws from the last E step of the EM algorithm}
}
\value{
a list made of the following components:
\itemize{
\item{u0}{a matrix of the Monte Carlo draws (from Rejection sampling if gibbs = F, or 
from Metropolis-within-Gibbs sampling if gibbs = T). Number rows = nMC, number columns = 
(number random effects)x(number groups). Organization of columns: first by random effect variable,
then by group within variable (i.e. Var1:Grp1 Var1:Grp2 ... Var1:GrpK Var2:Grp1 ... Varq:GrpK)}
\item{switch}{logical, did the sampling scheme switch from Rejection sampling to 
Metropolis-within-Gibbs sampling?}
}
}
\description{
\code{sample.mc2} calculates the Monte Carlo draws by either Rejection sampling or 
Metropolis-within-Gibbs sampling needed for Monte Carlo Expectation Conditional Minimization (MCECM)
}
