% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mc_functions.R
\name{sample_mc_adapt_BigMat}
\alias{sample_mc_adapt_BigMat}
\alias{sample_mc2_BigMat}
\title{Calculate Monte Carlo draws using Metropolis-within-Gibbs}
\usage{
sample_mc_adapt_BigMat(coef, ranef_idx, y, X, Z, nMC, trace = 0, family,
  group, d, okindex, uold, proposal_SD, batch, batch_length = 100,
  offset = 0, burnin_batchnum = 500)

sample_mc2_BigMat(coef, ranef_idx, y, X, Z, nMC, trace = 0, family,
  group, d, okindex, uold)
}
\arguments{
\item{coef}{a numeric vector of the fixed effects coefficients (from the latest EM iteration or
the final model of interest)}

\item{ranef_idx}{a vector of integers describing which random effects are non-zero (i.e. which
diagonal elements of the sigma matrix are non-zero)}

\item{y}{a numeric vector of the response variable}

\item{X}{a model matrix of the fixed effects covariates}

\item{Z}{a sparse model matrix of the random effects multiplied by the lower triangular cholesky
decomposition of the sigma matrix (from the latest EM iteration or the final model of interest)}

\item{nMC}{a positive integer for the initial number of Monte Carlo draws}

\item{trace}{an integer specifying print output to include as function runs. Default value is 0 
(details ...)}

\item{family}{a description of the error distribution and link function to be used in the model. 
For \code{fit_dat_B} this can be a character string naming a family function or a family function. 
See \code{family} options in the Details section.}

\item{group}{a factor vector of the grouping variable, converted to a factor of 
consecutive numeric integers}

\item{d}{integer, the number of groups present (number factors of the group vector)}

\item{okindex}{?}

\item{uold}{a matrix with a single row comprised of the last Monte Carlo draw from either the 
most recent E step of the EM algorithm or the final model}

\item{proposal_SD}{a matrix of dimension (number of groups)x(number of random effects) that
gives the proposal standard deviation for the adaptive random walk. The \code{glmmPen} MCEM 
algorithm initializes the proposal standard deviation as 1.0 for all variables and groups
and is updated at the start of every E step when the Metropolis-within-Gibbs adaptive random walk
algorithm is used.}

\item{batch}{integer specifying how many batches of posterior draws has already been sampled in
the MCEM algorithm. As the batch number increases, the proposal standard deviation for the adaptive 
random walk algorithm is adjusted by a smaller amount.}

\item{batch_length}{an integer specifying the number of posterior draws (ignoring thinning) used
to evaluate the acceptance rate of the random walk algorithm. This acceptance rate is then used
to adjust the proposal standard deviation if necessary.}

\item{burnin_batchnum}{an integer specifying the number of batches of (unthinned) posterior draws
to allow adjustments to the proposal standard deviation. For the MCEM algorithm, it is recommended
to allow for the adjustment of the proposal standard deviation at the beginning of each E step.}
}
\value{
a list made of the following components:
\item{u0}{list of the information needed to attached to a big.matrix object. Use 
\code{bigmemory::attach.big.matrix(u0)} to extract a big.matrix of the Monte Carlo draws of
the random effect posterior distribution. Number rows = nMC, number columns = 
(number random effects)*(number groups) = ncol(Z). Organization of columns: first by random effect 
variable, then by group within variable (i.e. Var1:Grp1 Var1:Grp2 ... Var1:GrpK Var2:Grp1 ... Varq:GrpK)}
\item{gibbs_accept_rate}{matrix of dimension (number groups)x(number random effects). Each entry
gives the proportion of accepted draws for the particular random effect variable in group k 
either in the entire \code{nMC} draws (for \code{sample_mc2_BigMat}) or for the last 
\code{batch_length} draws.}
\item{proposal_SD}{a matrix of dimension (number of groups)x(number of random effects) that is
only output from \code{sample_mc_adapt_BigMat}. It records the updated proposal standard deviation
for the particular random effect variable in group k.}
\item{updated_batch}{integer specifying the updated number of batches of posterior draws that 
have been sampled.}
}
\description{
Samples draws from the posterior distribution of the random effects using 
Metropolis-within-Gibbs. \code{sample_mc_adapt_BigMat} samples using adaptive 
random walk and \code{sample_mc2_BigMat} samples using independence sampling. See
\code{\link{adaptControl}} for possible adaptive random walk adjustments. Draws are automatically
thinned to only record every fifth sample.
}
