% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_mc_inner}
\alias{sample_mc_inner}
\title{Obtaining Rejection Sampling Monte Carlo Draws}
\usage{
sample_mc_inner(f, z, y, t, NMC, trace)
}
\arguments{
\item{f}{a one-column matrix of the fixed effects component of the linear predictor for 
the individuals/observations in a specific group (i.e. group k)}

\item{z}{a sub-matrix of the full Z matrix (see output information of \code{\link{formulaData}} 
for details); the columns pertain to the columns of Z associated 
with a particular group, the rows peratin to the rows of Z that are associated with
the group of interest}

\item{y}{the sub-vector of the original response vector that is associated with the particular
group of interest}

\item{NMC}{positive integer, number of Monte Carlo draws to obtain}

\item{trace}{an integer specifying print output to include as function runs. If trace = 2, then 
will output acceptance rate information}
}
\value{
A matrix of Rejection sampling Monte Carlo draws. nrows: \code{NMC}, columns: 
each column pertains to one of the q random effects variables that are associated with the 
group of interest (ncols = q). If the algorithm errored-out due to an unacceptably low 
acceptance rate, then a matrix of all zeros of dimension (NMC-3)xq is output.
}
\description{
Description
}
