% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_options.R
\name{adaptControl}
\alias{adaptControl}
\title{Control of Metropolis-within-Gibbs Adaptive Random Walk Sampling Procedure

Constructs control structure for the adaptive random walk Metropolis-within-Gibbs sampling procedure.}
\usage{
adaptControl(batch_length = 100, offset = 0, burnin_batchnum = 500)
}
\arguments{
\item{batch_length}{positive integer specifying the number of posterior draws to collect
before the proposal variance is adjusted based on the acceptance rate of the last 
\code{batch_length} posterior draws}

\item{offset}{positive integer specifying an offset value for the increment of the proposal
variance adjustment. Used to ensure the required diminishing adaptation condition. Default 8500.
\verb{
increment = min(0.01, 1 / sqrt(batch*batch_length + offset))
}}

\item{burnin_batchnum}{positive integer specifying the maximum number of batches (of length
\code{batch_length}) during which the proposal variance can be adjusted. Default 1,000.}

\item{nMC_burnin}{positive integer specifying teh number of posterior draws to use as
burnin for each E step in the EM algorithm. Default 200. Function will not allow \code{nMC_burnin}
to be less than 100.}
}
\value{
The *Control functions return a list (inheriting from class "\code{adaptControl}") 
containing parameter specifications for the adaptive random walk sampling procedure.
}
\description{
Control of Metropolis-within-Gibbs Adaptive Random Walk Sampling Procedure

Constructs control structure for the adaptive random walk Metropolis-within-Gibbs sampling procedure.
}
