% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dat.R
\name{fit_dat}
\alias{fit_dat}
\title{Fit a penalized generalized mixed model (pGLMM) via Monte Carlo Expectation Conditional 
Minimization (MCECM)}
\usage{
fit_dat(dat, lambda0 = 0, lambda1 = 0, conv = 0.001, nMC = 1000,
  family = "binomial", trace = 0, penalty = "grMCP", d = 2, group,
  alpha = 1, nMC_max = 5000, returnMC = F, ufull = NULL,
  coeffull = NULL, gibbs = T, maxitEM = 100, ufullinit = NULL)
}
\arguments{
\item{dat}{Input list object specifying y (response vector), X (matrix of all covariates), 
Z (matrix of covariates with possible random effects), and group (vector whose value indicates 
the study, batch, or other group identity to which on observation belongs), with each row an 
observation}
}
\description{
Fit a penalized generalized mixed model (pGLMM) via Monte Carlo Expectation Conditional 
Minimization (MCECM)
}
