% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dat_MstepB.R
\name{fit_dat_B}
\alias{fit_dat_B}
\title{Fit a Penalized Generalized Mixed Model via Monte Carlo Expectation Conditional 
Minimization (MCECM)}
\usage{
fit_dat_B(dat, lambda0 = 0, lambda1 = 0, conv = 0.001,
  family = "binomial", offset_fit = NULL, trace = 0,
  penalty = c("MCP", "SCAD", "lasso"), alpha = 1,
  gamma_penalty = switch(penalty, SCAD = 4, 3),
  group_X = 0:(ncol(dat$X) - 1), nMC = 1000, nMC_max = 10000,
  t = 2, returnMC = T, ufull = NULL, coeffull = NULL,
  ufullinit = NULL, maxitEM = 100, maxit_CD = 250, M = 10^4,
  gibbs = T, MwG_sampler = c("random_walk", "independence"),
  adapt_RW_options = adaptControl(), covar = c("unstructured",
  "independent"), fit_type = 1)
}
\arguments{
\item{dat}{a list object specifying y (response vector), X (model matrix of all covariates), 
Z (model matrix for the random effects), and group (vector whose value indicates 
the study, batch, or other group identity to which on observation belongs), with each row an 
observation}

\item{lambda0}{a non-negative numeric penalty parameter for the fixed effects parameters}

\item{lambda1}{a non-negative numeric penalty parameter for the grouped random effects
covariance parameters}

\item{conv}{a non-negative numeric convergence criteria for the convergence of the 
log likelihood in the EM algorithm. Default is 0.001.}

\item{family}{a description of the error distribution and link function to be used in the model. 
For \code{fit_dat_B} this can be a character string naming a family function or a family function. 
See \code{family} options in the Details section.}

\item{offset_fit}{This can be used to specify an a priori known component to be included in the 
linear predictor during fitting. This should be NULL or a numeric vector of length equal to the 
number of cases.}

\item{trace}{an integer specifying print output to include as function runs. Default value is 0 
(details ...)}

\item{penalty}{character descripting the type of penalty to use in the variable selection procedure.
Options include 'MCP', 'SCAD', and 'lasso'. Default is MCP penalty.}

\item{alpha}{Tuning parameter for the Mnet estimator which controls the relative contributions 
from the MCP/SCAD/lasso penalty and the ridge, or L2, penalty. \code{alpha=1} is equivalent to 
the MCP/SCAD/lasso penalty, while \code{alpha=0} is equivalent to ridge regression. However,
\code{alpha=0} is not supported; \code{alpha} may be arbibrarily small, but not exactly zero}

\item{gamma_penalty}{The tuning parameter of the MCP and SCAD penalties. Not used by Lasso penalty.
Default is 4.0 for SCAD and 3.0 for MCP.}

\item{nMC}{a positive integer for the initial number of Monte Carlo draws}

\item{nMC_max}{a positive integer for the maximum number of allowed Monte Carlo draws}

\item{returnMC}{logical, should the final Monte Carlo draws be returned? Default \code{TRUE}.}

\item{maxitEM}{a positive integer for the maximum number of allowed EM iterations}

\item{maxit_CD}{a positive integer for the maximum number of allowed interations for the
coordinate descent algorithms used within each EM iteration.}

\item{gibbs}{logical, should Metropolis-within-Gibbs sampling be used for the Monte Carlo draws 
(if \code{TRUE}) or should Rejection sampling be used for the draws (if \code{FALSE})?}
}
\value{
a list with the following elements:
\item{coef}{a numeric vector of coefficients of fixed effects estimates and 
non-zero estimates of the lower-triangular cholesky decomposition of the random effects
covariance matrix (in vector form)}
\item{sigma}{random effects covariance matrix}
\item{lambda0, lambda1}{the penalty parameters input into the function}
\item{covgroup}{?}
\item{J}{a sparse matrix of dimension q^2 x (q(q+1)/2) (where q = number of random effects) that 
transforms the non-zero elements of the lower-triangular cholesky decomposition of the random 
effects covariance matrix into a vector}
\item{ll}{estimate of the log likelihood, calculated using the Pajor method}
\item{BICh}{the hybrid BIC estimate described in Delattre, Lavielle, and Poursat (2014)}
\item{u}{a matrix of the Monte Carlo draws. Organization of columns: first by random effect variable,
then by group within variable (i.e. Var1:Grp1 Var1:Grp2 ... Var1:GrpK Var2:Grp1 ... Varq:GrpK)
Output if \code{returnMC} = \code{TRUE}}
\item{rej_to_gibbs}{logical, did the Monte Carlo sampling switch from Rejection sampling 
to Metropolis-within-Gibbs sampling due to unacceptably small acceptance rates in the Rejection sampling?
Output only if started initially with Rejection sampling}
}
\description{
Description
}
\section{Details}{
 
Accepted families: binomial, poisson, gaussian. Currently, the algorithm is only equipted to handle
the canonical links associated with these families (logit link for binomial, log link for poisson,
identity link for gaussian).
}

