% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmPen.R
\name{glmmPen}
\alias{glmmPen}
\title{Fit Penalized Generalized Mixed Models via Monte Carlo Expectation Conditional 
Minimization (MCECM)}
\usage{
glmmPen(
  formula,
  data = NULL,
  family = "binomial",
  covar = c("unstructured", "independent"),
  offset = NULL,
  na.action = na.omit,
  fixef_noPen = NULL,
  penalty = c("MCP", "SCAD", "lasso"),
  alpha = 1,
  gamma_penalty = switch(penalty[1], SCAD = 4, 3),
  optim_options = optimControl(),
  adapt_RW_options = adaptControl(),
  trace = 0,
  tuning_options = selectControl(),
  BICq_posterior = NULL
)
}
\arguments{
\item{formula}{a two-sided linear formula object describing both the fixed effects and 
random effects part of the model, with the response on the left of a ~ operator and the terms, 
sepearated by + operators, on the right. Random-effects terms are distinguished by vertical bars 
("|") separating expression for design matrices from the grouping factor. \code{formula} should be 
of the same format needed for \code{\link[lme4]{glmer}} in package \pkg{lme4}. Only one grouping factor 
will be recognized. The random effects covariates need to be a subset of the fixed effects covariates.
The offset must be specified outside of the formula in the 'offset' argument.}

\item{data}{an optional data frame containing the variables named in \code{formula}. If \code{data} is 
omitted, variables will be taken from the environment of \code{formula} (if specified as a formula).}

\item{family}{a description of the error distribution and link function to be used in the model. 
Currently, the \code{glmmPen} algorithm allows the binomial, gaussian, and poisson families
with canonical links only.}

\item{covar}{character string specifying whether the covariance matrix should be unstructured
("unstructured") or diagonal with no covariances between variables ("independent").
Default is "unstructured", but if the number of random effects (not including the intercept) is 
greater than or equal to 10 (i.e. high dimensional), then the algorithm automatically assumes an 
independent covariance structure and \code{covar} is switched to "independent".}

\item{offset}{This can be used to specify an \emph{a priori} known component to be included in the 
linear predictor during fitting. Default set to \code{NULL} (no offset). If the data 
argument is \code{NULL}, this should be a numeric vector of length equal to the 
number of cases (the response). If the data argument specifies a data.frame, the offset
argument should specify the name of a column in the data.frame.}

\item{na.action}{a function that indicates what should happen when the data contain NAs. Only the 
option \code{na.omit} are recognized by this function.}

\item{fixef_noPen}{Optional vector of 0's and 1's of the same length as the number of fixed effects covariates
used in the model. Value 0 indicates the variable should not have its fixed effect coefficient
penalized, 1 indicates that it can be penalized. Order should correspond to the same order of the 
fixed effects given in the formula.}

\item{penalty}{character describing the type of penalty to use in the variable selection procedure.
Options include 'MCP', 'SCAD', and 'lasso'. Default is MCP penalty. If the random effect covariance
matrix is "unstructured", then a group MCP, group SCAD, or group Lasso penalty is used on the 
random effects coefficients.}

\item{alpha}{Tuning parameter for the Mnet estimator which controls the relative contributions 
from the MCP/SCAD/lasso penalty and the ridge, or L2, penalty. \code{alpha=1} is equivalent to 
the MCP/SCAD/lasso penalty, while \code{alpha=0} is equivalent to ridge regression. However,
\code{alpha=0} is not supported; \code{alpha} may be arbibrarily small, but not exactly zero}

\item{gamma_penalty}{The tuning parameter of the MCP and SCAD penalties. Not used by Lasso penalty.
Default is 4.0 for SCAD and 3.0 for MCP.}

\item{optim_options}{a structure of class "optimControl" created 
from function \code{\link{optimControl}} that specifies optimization parameters. See the 
documentation for \code{\link{optimControl}} for more details on defaults.}

\item{adapt_RW_options}{a list of class "adaptControl" from function \code{\link{adaptControl}} 
containing the control parameters for the adaptive random walk Metropolis-within-Gibbs procedure. 
Ignored if \code{\link{optimControl}} parameter \code{sampler} is set to "stan" (default) or "independence".}

\item{trace}{an integer specifying print output to include as function runs. Default value is 0. 
See Details for more information about output provided when trace = 0, 1, or 2.}

\item{tuning_options}{a list of class selectControl or lambdaControl resulting from 
\code{\link{selectControl}} or \code{\link{lambdaControl}} containing additional control parameters.
When function \code{glmm} is used,the algorithm may be run using one specific set of
penalty parameters \code{lambda0} and \code{lambda1} by specifying such values in \code{lambdaControl()}. 
The default for \code{glmm} is to run the model fit with no penalization (\code{lambda0} = \code{lambda1} = 0).
When function \code{glmmPen} is run, \code{tuning_options} is specified usig \code{selectControl{}}. 
See the \code{\link{lambdaControl}} and \code{\link{selectControl}} documentation for further details.}

\item{BICq_posterior}{an optional character string expressing the path and file name of a text file that 
the will store or currently stores a \code{big.matrix} of the posterior draws from the full model.
These full model posterior draws will be used in BIC-ICQ calculations if these calculations
are requested. The name of the file should include a .txt extension. If this argument is
specified as \code{NULL} (default) and BIC-ICQ calculations are requested, the posterior draws
will be saved in the file 'BICq_Posterior_Draws.txt' in the working directory.
See 'Details' section for additional details.}
}
\value{
A reference class object of class \code{\link{pglmmObj}} for which many methods are 
available (e.g. \code{methods(class = "pglmmObj")})
}
\description{
\code{glmmPen} is used to fit penalized generalized mixed models via Monte Carlo Expectation 
Conditional Minimization (MCECM) and select the best model using BIC-type selection criteria
}
\details{
\code{BICq_posterior}: If the \code{BIC_option} in \code{\link{selectControl}} (\code{tuning_options}) is specified 
to be 'BICq', this requests the calculation of the BIC-ICQ criterion during the selection
process. For the BIC-ICQ criterion to be calculated, a full model assuming a small valued 
lambda penalty needs to be fit, and the posterior draws from this full model need to be used. 
In order to avoid repetitive calculations of
this full model if secondary rounds of selection are desired, a \code{big.matrix} of these 
posterior draws will be saved in a .txt file. If the file already exits, the full model 
will not be fit again and the big.matrix of 
posterior draws will be read using \code{bigmemory::read.big.matrix}
(\code{bigmemory::read.big.matrix(filename, sep = ' ', type = 'double')}) and used in the BIC-ICQ 
calcuations. If the file does not exist or \code{BICq_posterior} 
is specified as \code{NULL}, the full model will be fit and the full model posterior
draws will be saved at the specified file location (using \code{bigmemory::write.big.matrix})
or in the working directory with generic filename "BICq_Posterior_Draws.txt" (if \code{NULL} 
is specified). The algorithm will save 10^4 posterior draws automatically.

Trace details: The value of 0 outputs some general updates for each EM iteration (iteration number EM_iter,
number of MCMC draws nMC, average Euclidean distance between current coefficients and coefficients
from t iterations back EM_conv, and number of non-zero coefficients Non0 Coef). The value of 1
additionally outputs the updated coefficients, updated covariance matrix values, and the
number of coordinate descent iterations used for the M step for each
EM iteration. If Stan is not used as the E-step sampling mechanism, 
the value of 2 outputs all of the above plus gibbs acceptance rate information
for the adaptive random walk and independence samplers and the updated proposal standard deviation
for the adaptive random walk.
}
