% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmPen.R
\name{glmmPen}
\alias{glmmPen}
\title{Fit Penalized Generalized Mixed Models via Monte Carlo Expectation Conditional 
Minimization (MCECM)}
\usage{
glmmPen(
  formula,
  data = NULL,
  family = "binomial",
  covar = NULL,
  offset = NULL,
  na.action = na.omit,
  fixef_noPen = NULL,
  penalty = c("MCP", "SCAD", "lasso"),
  alpha = 1,
  gamma_penalty = switch(penalty[1], SCAD = 4, 3),
  optim_options = optimControl(),
  adapt_RW_options = adaptControl(),
  trace = 0,
  tuning_options = selectControl(),
  BICq_posterior = NULL
)
}
\arguments{
\item{formula}{a two-sided linear formula object describing both the fixed effects and 
random effects part of the model, with the response on the left of a ~ operator and the terms, 
sepearated by + operators, on the right. Random-effects terms are distinguished by vertical bars 
("|") separating expression for design matrices from the grouping factor. \code{formula} should be 
of the same format needed for \code{\link[lme4]{glmer}} in package \pkg{lme4}. Only one grouping factor 
will be recognized. The random effects covariates need to be a subset of the fixed effects covariates.
The offset must be specified outside of the formula in the 'offset' argument.}

\item{data}{an optional data frame containing the variables named in \code{formula}. If \code{data} is 
omitted, variables will be taken from the environment of \code{formula} (if specified as a formula).}

\item{family}{a description of the error distribution and link function to be used in the model. 
Currently, the \code{glmmPen} algorithm allows the binomial, gaussian, and poisson families
with canonical links only.}

\item{covar}{character string specifying whether the covariance matrix should be unstructured
("unstructured") or diagonal with no covariances between variables ("independent").
Default is set to \code{NULL}. If \code{covar} is set to \code{NULL} and the number of random effects
predictors (not including the intercept) is 
greater than or equal to 10 (i.e. high dimensional), then the algorithm automatically assumes an 
independent covariance structure and \code{covar} is set to "independent". Otherwise if \code{covar}
is set to \code{NULL} and the number of random effects predictors is less than 10, then the
algorithm automatically assumes an unstructured covariance structure and \code{covar} is set to "unstructured".}

\item{offset}{This can be used to specify an \emph{a priori} known component to be included in the 
linear predictor during fitting. Default set to \code{NULL} (no offset). If the data 
argument is \code{NULL}, this should be a numeric vector of length equal to the 
number of cases (the response). If the data argument specifies a data.frame, the offset
argument should specify the name of a column in the data.frame.}

\item{na.action}{a function that indicates what should happen when the data contain NAs. Only the 
option \code{na.omit} are recognized by this function.}

\item{fixef_noPen}{Optional vector of 0's and 1's of the same length as the number of fixed effects covariates
used in the model. Value 0 indicates the variable should not have its fixed effect coefficient
penalized, 1 indicates that it can be penalized. Order should correspond to the same order of the 
fixed effects given in the formula.}

\item{penalty}{character describing the type of penalty to use in the variable selection procedure.
Options include 'MCP', 'SCAD', and 'lasso'. Default is MCP penalty. If the random effect covariance
matrix is "unstructured", then a group MCP, group SCAD, or group Lasso penalty is used on the 
random effects coefficients.}

\item{alpha}{Tuning parameter for the Mnet estimator which controls the relative contributions 
from the MCP/SCAD/lasso penalty and the ridge, or L2, penalty. \code{alpha=1} is equivalent to 
the MCP/SCAD/lasso penalty, while \code{alpha=0} is equivalent to ridge regression. However,
\code{alpha=0} is not supported; \code{alpha} may be arbibrarily small, but not exactly zero}

\item{gamma_penalty}{The tuning parameter of the MCP and SCAD penalties. Not used by Lasso penalty.
Default is 4.0 for SCAD and 3.0 for MCP.}

\item{optim_options}{a structure of class "optimControl" created 
from function \code{\link{optimControl}} that specifies optimization parameters. See the 
documentation for \code{\link{optimControl}} for more details on defaults.}

\item{adapt_RW_options}{a list of class "adaptControl" from function \code{\link{adaptControl}} 
containing the control parameters for the adaptive random walk Metropolis-within-Gibbs procedure. 
Ignored if \code{\link{optimControl}} parameter \code{sampler} is set to "stan" (default) or "independence".}

\item{trace}{an integer specifying print output to include as function runs. Default value is 0. 
See Details for more information about output provided when trace = 0, 1, or 2.}

\item{tuning_options}{a list of class selectControl or lambdaControl resulting from 
\code{\link{selectControl}} or \code{\link{lambdaControl}} containing additional control parameters.
When function \code{glmm} is used,the algorithm may be run using one specific set of
penalty parameters \code{lambda0} and \code{lambda1} by specifying such values in \code{lambdaControl()}. 
The default for \code{glmm} is to run the model fit with no penalization (\code{lambda0} = \code{lambda1} = 0).
When function \code{glmmPen} is run, \code{tuning_options} is specified usig \code{selectControl{}}. 
See the \code{\link{lambdaControl}} and \code{\link{selectControl}} documentation for further details.}

\item{BICq_posterior}{an optional character string expressing the path and file basename of a file combination that 
the will file-back or currently file-backs a \code{big.matrix} of the posterior draws from the full model.
These full model posterior draws will be used in BIC-ICQ calculations if these calculations
are requested. If this argument is
specified as \code{NULL} (default) and BIC-ICQ calculations are requested, the posterior draws
will be saved in the file combination 'BICq_Posterior_Draws.bin' and 'BICq_Posterior_Draws.desc'
in the working directory.
See 'Details' section for additional details about the required format of \code{BICq_posterior}
and the file-backed big matrix.}
}
\value{
A reference class object of class \code{\link{pglmmObj}} for which many methods are 
available (e.g. \code{methods(class = "pglmmObj")})
}
\description{
\code{glmmPen} is used to fit penalized generalized mixed models via Monte Carlo Expectation 
Conditional Minimization (MCECM) and select the best model using BIC-type selection criteria
}
\details{

}
