% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmPen.R
\name{glmmPen}
\alias{glmmPen}
\title{Fit a Penalized Generalized Mixed Model via Monte Carlo Expectation Conditional 
Minimization (MCECM)}
\usage{
glmmPen(formula, data = NULL, family = "binomial",
  na.action = na.omit, offset = NULL, fixef_noPen = NULL,
  penalty = c("MCP", "SCAD", "lasso"), alpha = 1,
  gamma_penalty = switch(penalty[1], SCAD = 4, 3),
  optim_options = optimControl(), adapt_RW_options = adaptControl(),
  returnMC = T, trace = 0, tuning_options = selectControl())
}
\arguments{
\item{formula}{a two-sided linear formula object describing both the fixed-effects and 
random-effects part of the model, with the response on the left of a ~ operator and the terms, 
sepearated by + operators, on the right. Random-effects terms are distinguished by vertical bars 
("|") separating expression for design matrices from grouping factors. (\code{formula} should be 
of the same format needed for \code{\link[lme4]{glmer}} in package \pkg{lme4}.) Only one grouping factor 
will be recognized. The random effects covariates need to be a subset of the fixed effects covariates.}

\item{data}{an optional data frame containing the variables named in \code{formula}. Although 
\code{data} is optional, the package authors \emph{strongly} recommend its use. If \code{data} is 
omitted, variables will be taken from the environment of \code{formula} (if specified as a formula).}

\item{family}{a description of the error distribution and link function to be used in the model. 
For \code{fit_dat_B} this can be a character string naming a family function or a family function. 
See \code{family} options in the Details section.}

\item{na.action}{a function that indicates what should happen when the data contain NAs. Only the 
options \code{na.omit} and \code{na.pass} are recognized by this function.}

\item{offset}{This can be used to specify an \emph{a priori} known component to be included in the 
linear predictor during fitting. This should be \code{NULL} or a numeric vector of length equal to the 
number of cases. Currently, the formula does not allow specification of an offset.}

\item{fixef_noPen}{Optional vector of 0's and 1's of the same length as the number of fixed effect covariates
used in the model. Value 0 indicates the variable should not have its fixed effect coefficient
penalized, 1 indicates that it can be penalized. Order should correspond to the same order of the 
fixed effects given in the formula.}

\item{penalty}{character descripting the type of penalty to use in the variable selection procedure.
Options include 'MCP', 'SCAD', and 'lasso'. Default is MCP penalty.}

\item{alpha}{Tuning parameter for the Mnet estimator which controls the relative contributions 
from the MCP/SCAD/lasso penalty and the ridge, or L2, penalty. \code{alpha=1} is equivalent to 
the MCP/SCAD/lasso penalty, while \code{alpha=0} is equivalent to ridge regression. However,
\code{alpha=0} is not supported; \code{alpha} may be arbibrarily small, but not exactly zero}

\item{gamma_penalty}{The tuning parameter of the MCP and SCAD penalties. Not used by Lasso penalty.
Default is 4.0 for SCAD and 3.0 for MCP.}

\item{optim_options}{a list of class "optimControl" created from function \code{\link{optimControl}}
that specifies optimization parameters.}

\item{adapt_RW_options}{a list of class "adaptControl" from function \code{\link{adaptControl}} 
containing the control parameters for the adaptive random walk Metropolis-within-Gibbs procedure. 
Ignored if \code{\link{optimControl}} parameter \code{MwG_sampler} is set to "independence"}

\item{returnMC}{logical, should \code{nMC_report} Monte Carlo draws be returned? Default \code{TRUE}.}

\item{trace}{an integer specifying print output to include as function runs. Default value is 0 
(details ...)}

\item{tuning_options}{a list of class selectControl or lambdaControl resulting from 
\code{\link{selectControl}} or \code{\link{lambdaControl}} containing additional control parameters.
If the user wants to run the algorithm using one specific set of
penalty parameters \code{lambda0} and \code{lambda1}, then use \code{lambdaControl()}. 
If no penalization is desired, use this setting with \code{lambda0} = \code{lambda1} = 0.
If the user wants to run the algorithm over multiple possible \code{lambda0} and \code{lambda1},
then use \code{selectControl{}}. See the \code{\link{lambdaControl}} and \code{\link{selectControl}}
documentation for details}
}
\value{
An reference class object of class \code{\link{pglmmObj}} for which many methods are 
available (e.g. \code{methods(class = "pglmmObj")})
}
\description{
\code{glmmPen} is used to fit penalized generalized mixed models via Monte Carlo Expectation 
Conditional Minimization (MCECM)
}
