% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmPen.R
\name{glmmPen}
\alias{glmmPen}
\title{Fit a Penalized Generalized Mixed Model via Monte Carlo Expectation Conditional 
Minimization (MCECM)}
\usage{
glmmPen(formula, data = NULL, family = "binomial",
  na.action = na.omit, offset = NULL, weights = NULL,
  penalty = "grMCP", nMC = 100, nMC_max = 2000, returnMC = T,
  maxitEM = 100, trace = 0, conv = 0.001, alpha = 1,
  gibbs = NULL, control = lambdaControl())
}
\arguments{
\item{formula}{a two-sided linear formula object describing both the fixed-effects and 
random-effects part of the model, with the response on the left of a ~ operator and the terms, 
sepearated by + operators, on the right. Random-effects terms are distinguished by vertical bars 
("|") separating expression for design matrices from grouping factors. (\code{formula} should be 
of the same format needed for \code{\link[lme4]{glmer}} in package \pkg{lme4}.) Only one grouping factor 
will be recognized. The random effects covariates need to be a subset of the fixed effects covariates.}

\item{data}{an optional data frame containing the variables named in \code{formula}. Although 
\code{data} is optional, the package authors \emph{strongly} recommend its use. If \code{data} is 
omitted, variables will be taken from the environment of \code{formula} (if specified as a formula).}

\item{family}{a description of the error distribution and link function to be used in the model. 
For \code{fit_dat} this can be a character string naming a family function or a fmaily function. 
See \code{family} options in the Details section.}

\item{na.action}{a function that indicates what should happen when the data contain NAs. Only the 
options \code{na.omit} and \code{na.pass} are recognized by this function.}

\item{offset}{an optional vector of an \emph{a priori} known component to be included in the 
linear predictor during fitting. One or more \code{\link[stats]{offset}} terms can be included in 
formula instead or as well. If more than one is specified, their sum is use. 
See \code{\link[stats]{model.offset}}.}

\item{weights}{an optional vector of 'prior weights' to be used in the fitting process.
Should be \code{NULL} or a numeric vector.}

\item{penalty}{character descripting the type of penalty to use in the variable selection procedure}

\item{nMC}{a positive integer for the initial number of Monte Carlo draws}

\item{nMC_max}{a positive integer for the maximum number of allowed Monte Carlo draws}

\item{returnMC}{logical, should the final Monte Carlo draws be returned? Default \code{TRUE}.}

\item{maxitEM}{a positive integer for the maximum number of allowed EM iterations}

\item{trace}{an integer specifying print output to include as function runs. Default value is 0 
(details ...), with alternative options of 1 (details ...) or 2 (output acceptance rate information
of the Monte Carlo draws computed in \code{\link{sample.mc2}})}

\item{conv}{a non-negative numeric convergence criteria for the convergence of the 
log likelihood in the EM algorithm}

\item{alpha}{?}

\item{gibbs}{logical, should Metropolis-within-Gibbs sampling be used for the Monte Carlo draws 
(if \code{TRUE}) or should Rejection sampling be used for the draws (if \code{FALSE})?}

\item{control}{a list (of correct class, resulting from lambdaControl() or selectControl()) 
containing control parameters. If the user wants to run the algorithm using one specific set of
penalty parameters \code{lambda0} and \code{lambda1}, then use \code{lambdaControl()}. 
If the user wants to run the algorithm over multiple possible \code{lambda0} and \code{lambda1},
then use \code{lambdaControl{}}. See the \code{\link{lambdaControl}} and \code{\link{selectControl}}
documentation for details}
}
\value{
An reference class object of class \code{\link{pglmmObj}} for which many methods are 
available (e.g. \code{methods(class = "pglmmObj")})
}
\description{
\code{glmmPen} is used to fit penalized generalized mixed models via Monte Carlo Expectation 
Conditional Minimization (MCECM)
}
