% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_options.R
\name{lambdaControl}
\alias{lambdaControl}
\alias{selectControl}
\title{Control of Penalization Parameters and Selection Criteria

Constructs control structures for penalized mixed model fitting.}
\usage{
lambdaControl(lambda0 = 0, lambda1 = 0)

selectControl(lambda0_seq = NULL, lambda1_seq = NULL, nlambda = 20,
  BIC_option = c("BICh", "BIC"))
}
\arguments{
\item{lambda0}{a non-negative numeric penalty parameter for the fixed effects parameters}

\item{lambda1}{a non-negative numeric penalty parameter for the grouped random effects
covariance parameters}

\item{lambda0_seq}{a range of non-negative numeric penalty parameter for the fixed 
effects parameters. If \code{NULL}, then a range will be calculated as described by (...).}

\item{lambda1_seq}{a range of non-negative numeric penalty parameter for the grouped random 
effects covariance parameters. If \code{NULL}, then a range will be calculated as described 
by (...).}

\item{nlambda}{positive integer specifying number of lambda with which to fit a model.
Ignored if \code{lambda0_seq} and \code{lambda1_seq} are specified by the user.}

\item{BIC_option}{character value specifing the selection criteria used to select the 'best' model.
Default "BICh" utilizes the hybrid BIC value described in Delattre, Lavielle, and Poursat (2014).
The regular "BIC" option penalty term uses (total non-zero coefficients)*(length(y) = total number observations)}
}
\value{
The *Control functions return a list (inheriting from class "\code{pglmmControl}") 
containing penalization parameter values, presented either as an individual set or as a range of
possible values.
}
\description{
Control of Penalization Parameters and Selection Criteria

Constructs control structures for penalized mixed model fitting.
}
