% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_options.R
\name{lambdaControl}
\alias{lambdaControl}
\alias{selectControl}
\title{Control of Penalization Parameters and Selection Criteria}
\usage{
lambdaControl(lambda0 = 0, lambda1 = 0)

selectControl(
  lambda0_seq = NULL,
  lambda1_seq = NULL,
  nlambda = 10,
  BIC_option = c("BICh", "BIC", "BICq", "BICNgrp")
)
}
\arguments{
\item{lambda0}{a non-negative numeric penalty parameter for the fixed effects parameters}

\item{lambda1}{a non-negative numeric penalty parameter for the (grouped) random effects
covariance parameters}

\item{lambda0_seq, lambda1_seq}{a range of non-negative numeric penalty parameters for the fixed 
and random effect parameters, respectively. If \code{NULL}, then a range will be automatically 
calculated. See 'Details' section for more details on these default calculations.}

\item{nlambda}{positive integer specifying number of lambda with which to fit a model.
Ignored if \code{lambda0_seq} and \code{lambda1_seq} are specified by the user.}

\item{BIC_option}{character string specifing the selection criteria used to select the 'best' model.
Default "BICh" utilizes the hybrid BIC value described in Delattre, Lavielle, and Poursat (2014).
The regular "BIC" option penalty term uses (total non-zero coefficients)*(length(y) = total number
observations). The "BICq" option specifies the BIC-ICQ criterion, which requires a fit of 
a full model; a small penalty is used for the fixed and/or random effects if the number of 
effects is 10 or greater.}
}
\value{
The *Control functions return a list (inheriting from class "\code{pglmmControl}") 
containing penalization parameter values, presented either as an individual set or as a range of
possible values.
}
\description{
Constructs control structures for penalized mixed model fitting.
}
\details{
If unspecified, the \code{lambda0_seq} and \code{lambda1_seq} numeric sequences are 
automatically calculated. For an initial
grid search using \code{\link{glmmPen}}, the range will be calculated in the same manner as 
\code{ncvreg} calculates the range: max penalizes all fixed and random effects to 0, min is a 
small portion of max, range is composed of \code{nlambda} values spread evenly on the log scale.
For secondary rounds run using \code{\link{glmmPen_FineSearch}}, the optimal lambda combination 
from the initial round (based on the specified \code{BIC_option}) will be used to calculate a 
finer grid search. The new max and min lambda values are the lambda values two 
positions away from the best lambda in the original lambda sequences for the fixed and random
effects.

The \code{lambda0} and \code{lambda1} arguments allow for a user to fit a model with a single 
non-zero penalty parameter combination. However, this is generally not recommended.
}
