% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_options.R
\name{optimControl}
\alias{optimControl}
\title{Control of Penalized Generalized Linear Mixed Model Fitting

Constructs the control structure for the optimization of the penalized mixed model fit algorithm.}
\usage{
optimControl(
  conv_EM = 0.001,
  conv_CD = 1e-04,
  nMC_burnin = 250,
  nMC_start = 1000,
  nMC_max = 10^4,
  nMC_report = 5000,
  maxitEM = 100,
  maxit_CD = 200,
  M = 10000,
  t = 2,
  mcc = 3,
  sampler = c("stan", "random_walk", "independence"),
  var_start = 1,
  max_cores = 1
)
}
\arguments{
\item{conv_EM}{a non-negative numeric convergence criteria for the convergence of the 
EM algorithm. Default is 0.001.}

\item{conv_CD}{a non-negative numeric convergence criteria for the convergence of the 
grouped coordinate descent loop within the M step of the EM algorithm. Default 0.0001.}

\item{nMC_burnin}{positive integer specifying the number of posterior draws to use as
burnin for each E step in the EM algorithm. Default 250. Function will not allow \code{nMC_burnin}
to be less than 100.}

\item{nMC_start}{a positive integer for the initial number of Monte Carlo draws.}

\item{nMC_max}{a positive integer for the maximum number of allowed Monte Carlo draws}

\item{nMC_report}{positive integer specifying number of Monte Carlo posterior draws to return
when the function ends. Default set to the minimum allowed value: 5,000. 
Warning: the returned draws are formatted as a regular
matrix (not a big.matrix). Therefore, depending on the number of random effect covariates (q)
and the number of groups (d), choose \code{nMC_report} such that a matrix of size 
\code{nMC_report} by (q*d) does not cause memory issues on your operating system.}

\item{maxitEM}{a positive integer for the maximum number of allowed EM iterations. Default equals 100.}

\item{maxit_CD}{a positive integer for the maximum number of allowed interations for the
coordinate descent algorithms used within each EM iteration. Default equals 250.}

\item{M}{positive integer specifying the number of posterior draws to use within the 
Pajor log-likelihood calculation}

\item{t}{the convergence criteria is based on the average Euclidean distance between 
the most recent coefficient estimate and the coefficient estimate from t EM iterations back.
Positive integer, default equals 2.}

\item{mcc}{the number of times the convergence critera must be met before the algorithm is
seen as having converged. (mcc for 'meet condition counter'). Default set to 3. Value retricted 
to be no less than 3.}

\item{sampler}{character string specifying whether the posterior draws of the random effects
should be drawn using Stan (default, from package rstan) or the Metropolis-within-Gibbs procedure 
incorporating an adaptive random walk sampler ("random_walk") or an
independence sampler ("independence").}

\item{var_start}{positive number specifying the starting values to initialize the variance
of the covariance matrix. Default set to 1.0.}
}
\value{
Function returns a list (inheriting from class "\code{optimControl}") 
containing fit and optimization criteria values used in optimization routine.
}
\description{
Control of Penalized Generalized Linear Mixed Model Fitting

Constructs the control structure for the optimization of the penalized mixed model fit algorithm.
}
