% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_options.R
\name{optimControl}
\alias{optimControl}
\title{Control of Penalized Generalized Linear Mixed Model Fitting

Constructs the control structure for penalized mixed model fitting.}
\usage{
optimControl(conv = 0.001, nMC_start = 1000, nMC_max = 10000,
  nMC_report = 5000, maxitEM = 100, maxit_CD = 250, M = 10000,
  t = 2, covar = c("unstructured", "independence"),
  MwG_sampler = c("random_walk", "independence"), gibbs = T,
  fit_type = 2)
}
\arguments{
\item{conv}{a non-negative numeric convergence criteria for the convergence of the 
log likelihood in the EM algorithm. Default is 0.001.}

\item{nMC_start}{a positive integer for the initial number of Monte Carlo draws.}

\item{nMC_max}{a positive integer for the maximum number of allowed Monte Carlo draws}

\item{nMC_report}{positive integer specifying number of Monte Carlo posterior draws to return
when the function ends. Default 5,000. Warning: the returned draws are formatted in a regular
matrix (not a big.matrix). Therefore, depending on the number of random effect covariates (q)
and the number of groups (d), choose \code{nMC_report} such that a matrix of size 
\code{nMC_report} by (q*d) does not cause memory issues on your operating system.}

\item{maxitEM}{a positive integer for the maximum number of allowed EM iterations. Default equals 100.}

\item{maxit_CD}{a positive integer for the maximum number of allowed interations for the
coordinate descent algorithms used within each EM iteration. Default equals 250.}

\item{M}{positive integer specifying the number of posterior draws to use within the 
Pajor log-likelihood calculation}

\item{t}{the convergence criteria is based on the average Euclidean distance between 
the most recent coefficient estimate and the coefficient estimate from t EM iterations back.
Positive integer, default equals 2.}

\item{covar}{character value specifying whether the covariance matrix should be unstructured
("unstructured") or diagonal with no covariances between variables ("independent").
Default is "unstructured", but if the number of random effects (including the intercept) is 
greater than 7 (i.e. high dimensional), then the algorithm automatically assumes an 
independent covariance structure (covar switched to "independent").}

\item{MwG_sampler}{character value specifying whether the Metropolis-within-Gibbs procedure 
should incorporate an adaptive random walk sampler (default, "random_walk") or an
independence sampler ("independence").}

\item{gibbs}{logical, should Metropolis-within-Gibbs sampling be used for the Monte Carlo draws 
(if \code{TRUE}) or should Rejection sampling be used for the draws (if \code{FALSE})? 
Default of gibbs = T. Gibbs = F not recommended if greater than 5 random effects}

\item{fit_type}{integer specifying the verion of the grouped coordinate descent algorithm to use.
Default equals 2, the 'naive fit', assuming that the random effects covariates do not need any
additional standardization or adjustments to account for absence of orthogonalization.
For the time being, this version is recommended. Version 4 adds standardization of the random
effects covariates compared to version 2. Version 1 adds an adjustment to account for absence
of orthogonalization, and version 3 adds standardization compared to version 1.}
}
\value{
The *Control functions return a list (inheriting from class "\code{optimControl}") 
containing fit and optimization criteria values used in optimization routine.
}
\description{
Control of Penalized Generalized Linear Mixed Model Fitting

Constructs the control structure for penalized mixed model fitting.
}
