% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_tune.R
\name{select_tune}
\alias{select_tune}
\title{Fit a Penalized Generalized Mixed Model via Monte Carlo Expectation Conditional 
Minimization (MCECM)}
\usage{
select_tune(dat, offset = NULL, family, group_X = 0:(ncol(dat$X) - 1),
  penalty, lambda0_range, lambda1_range, alpha = 1,
  gamma_penalty = switch(penalty[1], SCAD = 4, 3), returnMC = T,
  trace = 0, ufull = NULL, coeffull = NULL,
  adapt_RW_options = adaptControl(), optim_options = optimControl())
}
\arguments{
\item{dat}{a list object specifying y (response vector), X (model matrix of all covariates), 
Z (model matrix for the random effects), and group (vector whose value indicates 
the study, batch, or other group identity to which on observation belongs), with each row an 
observation}

\item{offset}{This can be used to specify an \emph{a priori} known component to be included in the 
linear predictor during fitting. This should be \code{NULL} or a numeric vector of length equal to the 
number of cases. Currently, the formula does not allow specification of an offset.}

\item{family}{a description of the error distribution and link function to be used in the model. 
For \code{fit_dat_B} this can be a character string naming a family function or a family function. 
See \code{family} options in the Details section.}

\item{penalty}{character descripting the type of penalty to use in the variable selection procedure.
Options include 'MCP', 'SCAD', and 'lasso'. Default is MCP penalty.}

\item{alpha}{Tuning parameter for the Mnet estimator which controls the relative contributions 
from the MCP/SCAD/lasso penalty and the ridge, or L2, penalty. \code{alpha=1} is equivalent to 
the MCP/SCAD/lasso penalty, while \code{alpha=0} is equivalent to ridge regression. However,
\code{alpha=0} is not supported; \code{alpha} may be arbibrarily small, but not exactly zero}

\item{gamma_penalty}{The tuning parameter of the MCP and SCAD penalties. Not used by Lasso penalty.
Default is 4.0 for SCAD and 3.0 for MCP.}

\item{returnMC}{logical, should \code{nMC_report} Monte Carlo draws be returned? Default \code{TRUE}.}

\item{trace}{an integer specifying print output to include as function runs. Default value is 0 
(details ...)}

\item{adapt_RW_options}{a list of class "adaptControl" from function \code{\link{adaptControl}} 
containing the control parameters for the adaptive random walk Metropolis-within-Gibbs procedure. 
Ignored if \code{\link{optimControl}} parameter \code{MwG_sampler} is set to "independence"}

\item{optim_options}{a list of class "optimControl" created from function \code{\link{optimControl}}
that specifies optimization parameters.}
}
\value{
A list with the following elements:
\item{results}{matrix of summary results for each lambda tuning parameter combination, used
to select the 'best' model}
\item{out}{list of \code{\link{fit_dat_B}} results for the models with the minimum BICh and 
minimum BIC values}
\item{coef}{matrix of coefficient results for each lambda tuning parameter combination. 
Rows correspond with the rows of the results matrix.}
}
\description{
\code{select_tune} is used to fit penalized generalized mixed models via Monte Carlo Expectation 
Conditional Minimization (MCECM) over multiple tuning parameters
}
